Added a few more TODOs
Created a method called full_propogation that is intedended to be run manually to clean up addition errors by starting at the very beginning.

Fixed a sign error when determining whether to add a new row to prevent deleting the last unfrozen row in the app_entries worksheet

Sorted the unresolved entries.

Changed a string comparison to a float comparison because it was having trouble with whole numbers vs floats.

Added a print statement to inform you that the program crashed.